project circularity
