+++ src/command/serve/cmd.ts
@@ -7,7 +7,8 @@
 
 import { Command } from "cliffy/command/mod.ts";
 
-import { kProjectType, projectContext } from "../../project/project-context.ts";
+import { kProjectType } from "../../project/project-shared.ts";
+import { projectContext } from "../../project/project-context.ts";
 import { projectType } from "../../project/types/project-types.ts";
 import { findOpenPort } from "./port.ts";
