+++ src/execute/observable/compile.ts
@@ -45,7 +45,7 @@ import {
 } from "../../config/constants.ts";
 
 import { RenderContext } from "../../command/render/types.ts";
-import { ProjectContext } from "../../project/project-context.ts";
+import { ProjectContext } from "../../project/project-shared.ts";
 import { languagesInMarkdown } from "../../core/jupyter/jupyter.ts";
 import { asHtmlId } from "../../core/html.ts";
 import { sessionTempFile } from "../../core/temp.ts";
