+++ src/project/project-create.ts
@@ -17,7 +17,7 @@ import { renderEjs } from "../core/ejs.ts";
 import { executionEngine } from "../execute/engine.ts";
 import { ExecutionEngine } from "../execute/types.ts";
 
-import { projectConfigFile } from "./project-context.ts";
+import { projectConfigFile } from "./project-shared.ts";
 import { ensureGitignore } from "./project-gitignore.ts";
 
 export interface ProjectCreateOptions {
