+++ src/project/project-shared.ts
@@ -1,0 +1,74 @@
+/*
+* project-shared.ts
+*
+* Copyright (C) 2020 by RStudio, PBC
+*
+*/
+
+import { existsSync } from "fs/exists.ts";
+import { join } from "path/mod.ts";
+import { PandocFlags } from "../config/flags.ts";
+import { Format, FormatExtras } from "../config/format.ts";
+
+export const kProjectType = "type";
+export const kProjectRender = "render";
+export const kProjectExecuteDir = "execute-dir";
+export const kProjectOutputDir = "output-dir";
+export const kProjectLibDir = "lib-dir";
+export const kProjectResources = "resources";
+
+export interface ProjectContext {
+  dir: string;
+  engines: string[];
+  files: {
+    input: string[];
+    resources?: string[];
+    config?: string[];
+    configResources?: string[];
+  };
+  config?: ProjectConfig;
+  formatExtras?: (
+    project: ProjectContext,
+    source: string,
+    flags: PandocFlags,
+    format: Format,
+  ) => Promise<FormatExtras>;
+}
+
+export interface ProjectConfig {
+  project: {
+    [kProjectType]?: string;
+    [kProjectRender]?: string[];
+    [kProjectExecuteDir]?: "file" | "project";
+    [kProjectOutputDir]?: string;
+    [kProjectLibDir]?: string;
+    [kProjectResources]?: string[];
+  };
+  [key: string]: unknown;
+}
+
+export function projectOutputDir(context: ProjectContext): string {
+  let outputDir = context.config?.project[kProjectOutputDir];
+  if (outputDir) {
+    outputDir = join(context.dir, outputDir);
+  } else {
+    outputDir = context.dir;
+  }
+  if (existsSync(outputDir)) {
+    return Deno.realPathSync(outputDir);
+  } else {
+    return outputDir;
+  }
+}
+
+export function projectConfigFile(dir: string): string | undefined {
+  return ["_quarto.yml", "_quarto.yaml"]
+    .map((file) => join(dir, file))
+    .find(existsSync);
+}
+
+export function projectVarsFile(dir: string): string | undefined {
+  return ["_variables.yml", "_variables.yaml"]
+    .map((file) => join(dir, file))
+    .find(existsSync);
+}
