+++ src/project/types/website/website-config.ts
@@ -14,7 +14,7 @@ import {
 } from "../../../config/constants.ts";
 import { isHtmlOutput } from "../../../config/format.ts";
 
-import { ProjectConfig } from "../../project-context.ts";
+import { ProjectConfig } from "../../project-shared.ts";
 
 export const kSite = "site";
