+++ src/project/types/website/website-search.ts
@@ -20,11 +20,8 @@ import { isHtmlContent } from "../../../core/mime.ts";
 import { FormatDependency } from "../../../config/format.ts";
 
 import { kBootstrapDependencyName } from "../../../format/html/format-html-shared.ts";
-import {
-  ProjectContext,
-  projectOffset,
-  projectOutputDir,
-} from "../../project-context.ts";
+import { ProjectContext, projectOutputDir } from "../../project-shared.ts";
+import { projectOffset } from "../../project-context.ts";
 import { ProjectOutputFile } from "../project-type.ts";
 
 import {
