+++ src/project/types/website/website.ts
@@ -13,11 +13,8 @@ import { resourcePath } from "../../../core/resources.ts";
 import { dirAndStem } from "../../../core/path.ts";
 import { isHtmlContent } from "../../../core/mime.ts";
 
-import {
-  ProjectContext,
-  projectOffset,
-  projectOutputDir,
-} from "../../project-context.ts";
+import { ProjectContext } from "../../project-shared.ts";
+import { projectOffset, projectOutputDir } from "../../project-context.ts";
 import { kProject404File } from "../../project-resources.ts";
 
 import {
